Development

+++ src/Modules/Blog/Explorer.Blog.API/Dtos/CommentDto.cs
@@ -8,9 +8,9 @@ namespace Explorer.Blog.API.Dtos
 {
     public class CommentDto
     {
-        public int Id { get; set; }
-        public int BlogId { get; set; }
-        public int UserId { get; set; }
+        public long Id { get; set; }
+        public long BlogId { get; set; }
+        public long UserId { get; set; }
         public DateTime CreationTime { get; set; }
         public DateTime LastModifiedTime { get; set; }
         public string Text { get; set; }
